fix(ArtyHub-Mislata/sprints#36): Arreglados los fallos de los tests de banco-back

+++ src/main/java/es/artyhub/banco_back/persistence/dao/jpa/MovimientoBancarioJpaDao.java
@@ -1,11 +1,12 @@
 package es.artyhub.banco_back.persistence.dao.jpa;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import es.artyhub.banco_back.persistence.dao.jpa.entity.MovimientoBancarioJpaEntity;
 
 public interface MovimientoBancarioJpaDao extends GenericJpaDao<MovimientoBancarioJpaEntity>{
-    MovimientoBancarioJpaEntity findByImporte(Long importe);
+    MovimientoBancarioJpaEntity findByImporte(BigDecimal importe);
     MovimientoBancarioJpaEntity findByConcepto(String concepto);
     List<MovimientoBancarioJpaEntity> findByCuentaId(Long cuenta_id);
     MovimientoBancarioJpaEntity insert(MovimientoBancarioJpaEntity movimientoBancarioJpaEntity, Long cuentaId);
